ElectronEdge/Project 3/electronedge proj3/medvis/backend/requirements.txt
# Web Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & AI
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.1

# LLM APIs
openai==1.3.0
anthropic==0.7.8

# Image Processing
opencv-python==4.8.1.78

# PDF Generation
reportlab==4.0.4

# HTTP Requests
requests==2.31.0

# Environment Management
python-dotenv==1.0.0

# Additional Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1